Eco_Learn-main/server/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pillow==10.4.0